move prettifying tests to workflows
