+++ tests/api/happy/test_epidoc_happy.py
@@ -2,17 +2,11 @@ from lxml import etree
 
 from pyepidoc.epidoc.epidoc import EpiDoc
 from pyepidoc.shared import head
-from pyepidoc.shared.testing import save_reload_and_compare_with_benchmark
 from pyepidoc.epidoc.dom import lang, line
 
 import pytest
 
 relative_filepaths = {
-    'ugly': 'tests/api/files/prettifying/ugly/ISic000552.xml',
-    'benchmark_lxml': 'tests/api/files/prettifying/benchmark/ISic000552_prettified_lxml.xml',
-    'benchmark_pyepidoc': 'tests/api/files/prettifying/benchmark/ISic000552_prettified_pyepidoc.xml',
-    'prettified_lxml': 'tests/api/files/prettifying/prettified/ISic000552_prettified_lxml.xml',
-    'prettified_pyepidoc': 'tests/api/files/prettifying/prettified/ISic000552_prettified_pyepidoc.xml',
     'ISic000001': 'tests/api/files/single_files_untokenized/ISic000001.xml',
     'ISic000552': 'tests/api/files/single_files_tokenized/ISic000552.xml',
     'persName_nested': 'tests/api/files/persName_nested.xml',
@@ -115,42 +109,6 @@ def test_load_relative_filepath_from_str(filepath:str):
 def test_materialclasses():
     doc = EpiDoc(relative_filepaths['ISic000001'])
     assert doc.materialclasses == ['#material.stone.marble']
-
-
-def test_prettify_doc_with_lxml():
-    """
-    Tests that the entire document is prettified correctly
-    using lxml's inbuilt prettifier.
-    Prettifies both the main document and the editions.
-    """
-
-    ugly = EpiDoc(relative_filepaths['ugly'])
-    prettified = ugly.prettify('lxml')
-    prettified.to_xml_file(relative_filepaths['prettified_lxml'])
-    prettified_str = etree.tostring(prettified._e)
-
-    benchmark = EpiDoc(relative_filepaths['benchmark_lxml'])
-    benchmark_str = etree.tostring(benchmark._e)
-
-    assert prettified_str == benchmark_str
-
-
-def test_prettify_doc_with_pyepidoc():
-
-    """
-    Tests that the entire document is prettified correctly
-    using pyepidoc's prettifier.
-    Prettifies both the main document and the editions.
-    """
-
-    ugly = EpiDoc(relative_filepaths['ugly'])
-    prettified = ugly.prettify('pyepidoc')
-
-    assert save_reload_and_compare_with_benchmark(
-        doc=prettified,
-        target_path=relative_filepaths['prettified_pyepidoc'],
-        benchmark_path=relative_filepaths['benchmark_pyepidoc']
-    )
 
 
 def test_punct():
